pyproject.toml
[project]
name = "huckleberry-api"
version = "0.1.10"
description = "Python API client for Huckleberry baby tracking app using Firebase Firestore"
readme = { file = "README.md", content-type = "text/markdown" }
license = { text = "MIT" }
authors = [
    { name = "Woyken", email = "13166802+Woyken@users.noreply.github.com" }
]
requires-python = ">=3.9"
dependencies = [
    "google-cloud-firestore>=2.11.0",
    "requests>=2.31.0",
]
keywords = ["huckleberry", "baby", "tracking", "firebase", "api"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.urls]
Homepage = "https://github.com/Woyken/huckleberry-homeassistant"
Repository = "https://github.com/Woyken/huckleberry-homeassistant"
Issues = "https://github.com/Woyken/huckleberry-homeassistant/issues"

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["uv_build>=0.9.13,<0.10.0"]
build-backend = "uv_build"
